third_party/WebKit/LayoutTests/http/tests/devtools/tracing/console-timeline-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests console.time and timeEnd methods.


Running: testStartStopTimeline
timestamp 0
one
timestamp 1
one
timestamp 2

Running: testStartStopMultiple
timestamp 0
one
timestamp 1
two
two
timestamp 2
one
timestamp 3
timestamp 4
timestamp 5

Running: testStartMultipleStopInsideEvals
timestamp 0
one
timestamp 1
two
timestamp 2
timestamp 3
two
timestamp 4
timestamp 5
one
timestamp 6 - FAIL

Running: testStopUnknown
timestamp 0
one
timestamp 1
timestamp 2
one
timestamp 3

Running: testStartFromPanel
timestamp 0
one
timestamp 1
one
timestamp 2

Running: testStopFromPanel
timestamp 0
one
timestamp 1
two
timestamp 2
